IMSI, CCID and module_rev read and insert into database added
Removed function pointer for manual or automatic serial number insert.
Added database_insert function for both.
IMSI, CCID and module revision numbers are read in the module_test function
with qftpc command, and inserted into the database along with the
serial number, hardware revision and product number of the device.